+++ gui/gui_functions.py
@@ -1,12 +1,15 @@
-from config import Colors as c
-from config import Padding as p
-from config import Spacer as sp
-from config import Text as t
-from config import Styles as st
-import utils
+from gui.gui_config import Colors as c
+from gui.gui_config import Padding as p
+from gui.gui_config import Spacer as sp
+from gui.gui_config import Text as t
+from gui.gui_config import Styles as st
+from yolo_processing.exercises import run_yolo
+
 import tkinter as tk
-import cv2
+import os
 import time
+import cv2
+
 
 selected_video = ''
 selected_exercise = 0
@@ -80,6 +83,19 @@ def record_new_video(root, source_frame):
     selected_video = 'exercise_videos/new_videos/recorded_video.mp4'
     create_exercise_frame(root, source_frame)
 
+def get_existing_videos():
+    folder_path = "assets/exercise_videos/existing_videos"
+    video_extensions = ['.mp4', '.avi', '.mkv', '.mov', '.wmv', '.flv', '.webm']
+    video_files = []
+
+    if os.path.exists(folder_path) and os.path.isdir(folder_path):
+        for root, dirs, files in os.walk(folder_path):
+            for file in files:
+                if any(file.endswith(ext) for ext in video_extensions):
+                    video_files.append(file)
+    else:
+        print("Folder does not exist or is not a directory.")
+    return video_files
 
 def create_source_frame(root):
     source_frame = tk.Frame(root, bg=c.background)
@@ -87,7 +103,7 @@ def create_source_frame(root):
     source_frame.pack(fill=tk.BOTH, expand=True)
     create_logo(source_frame, sp.source, 1)
 
-    options_list = utils.get_video_file_names()
+    options_list = get_existing_videos()
     selected_option = tk.StringVar(source_frame)
     selected_option.set("Izaberite video")
     option_menu = tk.OptionMenu(source_frame, selected_option, *options_list)
@@ -161,6 +177,7 @@ def create_exercise_frame(root, source_frame):
         exercise_label.bind("<Button-1>", lambda event, i=i: select_exercise(i, root, exercise_frame))   
 
 #Display
+
 def create_display_frame(root, exercise_frame):
     exercise_frame.destroy()
     print(selected_video)
@@ -186,11 +203,10 @@ def create_display_frame(root, exercise_frame):
         cap.set(cv2.CAP_PROP_POS_FRAMES, frame_counter)
         ret, frame = cap.read()
         if ret:
-            processed_frame = utils.process_frame(frame, selected_exercise)
-            utils.resize_frame(processed_frame)
+            run_yolo(frame, selected_exercise)
             cap.release()
 
-            image = tk.PhotoImage(file="assets/gui_images/display_frame.png")
+            image = tk.PhotoImage(file="assets/temporary_images/display_frame.png")
             display_label = tk.Label(vertical_frame, image=image, borderwidth=0, relief="flat")
             display_label.image = image
             display_label.grid(row=0, column=0)
