+++ yolo_processing/osoba.py
